move all custom packages under pkg
